Added loading states for many pages, error handling for review page updated

+++ src/components/reviews/delete.tsx
@@ -1,7 +1,18 @@
 import { supabase } from "@/lib/supabaseClient"
+import {useState} from "react"
+
 
 export default function DeleteCard({ id }: { id: string}) {
+    const [deleting, setDeleting] = useState(false);
+    const [loading, setLoading] = useState(false)
+
+
     const handleDelete = async () => { 
+        const confirmDelete = confirm("Are you sure you want to delete this review?")
+        if (!confirmDelete) return
+
+        setDeleting(true)
+
         const {error} = await supabase
         .from("reviews")
         .delete()
@@ -9,13 +20,20 @@ export default function DeleteCard({ id }: { id: string}) {
 
         if (error){
             console.error("Delete failed: ", error)
+            setDeleting(false)
         } else {
             //refresh page
             location.reload()
         }      
 }
 
+
 return (
-    <button onClick={handleDelete} className="bg-red-600 text-white px-3 py-1 rounded">Delete</button>
+    <button onClick={handleDelete}
+     disabled={deleting}
+     className="bg-red-600 text-white px-3 py-1 rounded">{deleting && (
+        <div className="animate-spin h-4 w-4 border-2 border-white border-t-transparent rounded-full"></div>
+      )}<span>{deleting ? "Deleting..." : "Delete Review"}</span>
+      </button>
     )
 }
